// nextjs/gbiz-next/src/packages/entrance/src/api/market-operation/types.gen.ts
/**
      * This file is auto-generated by @kucoin/api-generator
      * Owner: sean.shi@kupotech.com
      */

export type GenericResult_List_string__ = {
    code?: string;
    data?: Array<string>;
    msg?: string;
    retry?: boolean;
    success?: boolean;
};

export type GenericResult_boolean_ = {
    code?: string;
    data?: boolean;
    msg?: string;
    retry?: boolean;
    success?: boolean;
};

export type RecalledByEmailUsingPostData = {
    body?: never;
    path?: never;
    query?: {
        /**
         * 邮箱
         */
        email?: string;
        /**
         * 语言
         */
        language?: string;
        /**
         * 捕获方式
         */
        type?: number;
    };
    url: '/recalled-email';
};

export type RecalledByEmailUsingPostResponses = {
    /**
     * OK
     */
    200: GenericResult_boolean_;
};

export type RecalledByEmailUsingPostResponse = RecalledByEmailUsingPostResponses[keyof RecalledByEmailUsingPostResponses];

export type RecalledByPhoneNumberUsingPostData = {
    body?: never;
    path?: never;
    query?: {
        /**
         * 国际区号
         */
        countryCode?: string;
        /**
         * 语言
         */
        language?: string;
        /**
         * 电话号码
         */
        phone?: string;
        /**
         * 捕获方式
         */
        type?: number;
    };
    url: '/recalled-phone';
};

export type RecalledByPhoneNumberUsingPostResponses = {
    /**
     * OK
     */
    200: GenericResult_boolean_;
};

export type RecalledByPhoneNumberUsingPostResponse = RecalledByPhoneNumberUsingPostResponses[keyof RecalledByPhoneNumberUsingPostResponses];

export type GetMailSuffixesUsingGetData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/mail-suffixes';
};

export type GetMailSuffixesUsingGetResponses = {
    /**
     * OK
     */
    200: GenericResult_List_string__;
};

export type GetMailSuffixesUsingGetResponse = GetMailSuffixesUsingGetResponses[keyof GetMailSuffixesUsingGetResponses];

export type ClientOptions = { baseURL: string; };